backend/requirements.txt
fastapi
async-fastapi-jwt-auth
fastapi-jwt-auth
attrs
bcrypt
cmake
uvicorn
fastapi-cli==0.0.4
h11
huggingface-hub
iniconfig
lxml
numpy
openai
pandas
pubmed-parser
PyJWT
python-dotenv==1.0.1
python-multipart
PyYAML
regex
safetensors
scipy
sentencepiece
spacy
tiktoken
tokenizers
torch
tqdm
transformers
typer
typing_extensions
urllib3
yarl
zstandard
python-docx
python-pptx
PyPDF2
opensearch-py
qdrant-client
sentence-transformers
peft
nltk
asyncpg
bitsandbytes
psycopg2
langchain
langchain_community
langchain_qdrant
langchain_text_splitters
python-dateutil